# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления заявками на системную нагрузку",
        "title": "System Load API",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8080",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/strategies": {
            "get": {
                "description": "Возвращает все стратегии или фильтрует по названию",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strategies"
                ],
                "summary": "Получить список стратегий",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название стратегии для поиска",
                        "name": "Title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список стратегий",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/web_backend_internal_app_serializer.ShardingStrategyJSON"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новую стратегию шардинга (с возможной загрузкой изображения/видео)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strategies"
                ],
                "summary": "Создать стратегию",
                "parameters": [
                    {
                        "description": "Данные новой стратегии",
                        "name": "strategy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web_backend_internal_app_serializer.ShardingStrategyJSON"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданная стратегия",
                        "schema": {
                            "$ref": "#/definitions/web_backend_internal_app_serializer.ShardingStrategyJSON"
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/strategies/{id}": {
            "get": {
                "description": "Возвращает информацию о стратегии по идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strategies"
                ],
                "summary": "Получить стратегию по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID стратегии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные стратегии",
                        "schema": {
                            "$ref": "#/definitions/web_backend_internal_app_serializer.ShardingStrategyJSON"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Стратегия не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "web_backend_internal_app_serializer.ShardingStrategyJSON": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "latency_coefficient": {
                    "type": "number"
                },
                "photo_url": {
                    "type": "string"
                },
                "reliability_coefficient": {
                    "type": "number"
                },
                "strategy_id": {
                    "type": "integer"
                },
                "throughput_coefficient": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}